rlcs-viz-app/requirements.txt
pandas
Pillow
plotly
numpy
scipy
shinywidgets
statsmodels
fastparquet